Rotate the picture direction
108/9/1 
Bob and I decided the direction of map.
So I need to change the map direction and send him the renew map (integer 2D array).
Although now I have already updated the map data, but this change has some problem.
The Exit and Start can't show on the map.
The positions' coordinates aren't intuitive. 
In the future these parts must be refactored.