The second version -- A* algorithm
Now it can find a path from start to goal. (1 start 1 goal)
And it will consider the obstacles and avoid them.
But now there is a problem that it calculates too long.
I tested from start to door1.
It takes about 133 seconds.
From start to door3.
It takes about 1994 seconds.